requirements/prod.txt
-r common.txt
dj-database-url==0.3.0
waitress==0.8.9
djangorestframework==3.1.3
django-cors-headers
